getTransactionCount
